cli: move CLI flags to a separate crate
In order to configure the execution of Wasmtime when run in the
Sightglass benchmark infrastructure, we need some way to alter the
various engine flags. This change resolves this by moving the parsing of
the common Wasmtime flags, `CommonOptions`, into its own crate.
